Corrected problem with pytorch

+++ mef/utils/pytorch/lighting/module.py
@@ -28,7 +28,7 @@ class _MefModel(pl.LightningModule, ABC):
     def _shared_step(self, batch, step_type):
         x, y = batch
 
-        preds = self._shared_step_output(x)
+        preds = self._shared_step_output(x).float()
 
         # preds is expected to in shape of [B] for binary and [B, C] for
         # multiclass
